Module jormungandr_automation::jormungandr

Re-exports

pub use self::explorer::compare_schema as compare_explorer_schema;

pub use self::explorer::Explorer;

pub use self::explorer::ExplorerError;

pub use self::explorer::ExplorerProcess;

Modules

explorer

grpc

Structs

BackwardCompatibleRest

Legacy tolerant rest api This layer returns raw strings without deserialization in order to assure compatibility and lack of serde errors

Block0ConfigurationBuilder

ConfiguredStarter

FaketimeConfig

JormungandrBootstrapper

JormungandrLogger

JormungandrParams

JormungandrProcess

JormungandrRest

Specialized rest api

JormungandrStateVerifier

LegacyNodeConfig

LegacyNodeConfigBuilder

Used to build configuration for legacy nodes. It uses yaml_rust instead of serde yaml serializer beacuse config model is always up to date with newest config schema while legacy node requires na old one

LegacyNodeConfigConverter

LegacyNodeConfigManager

MemPoolCheck

NodeConfigBuilder

NodeConfigManager

RawRest

struct intends to return raw reqwest response can be used to verify requests error codes or to poll until data is available

RemoteJormungandr

RemoteJormungandrBuilder

RestSettings

SecretModelFactory

Starter

Version

SemVer version as defined by https://semver.org.

Enums

EitherHashOrBlock0

FragmentNodeError

JormungandrError

LeadershipMode

LegacyConfigError

LogLevel

NodeBlock0

PersistenceMode

RestError

ShutdownError

StartupError

StartupVerificationMode

Status

TestingDirectory

Traits

ConfigurableNodeConfig

Abstraction over different versions of the node configuration.

FragmentNode

Functions

assert_accepted_rejected

assert_bad_request

download_last_n_releases

get_available_port

get_jormungandr_bin

uri_from_socket_addr

version_0_8_19

Type Definitions

NodeAlias