Struct jormungandr_automation::jormungandr::grpc::server::data::MockServerData

pub struct MockServerData {
    genesis_hash: Hash,
    protocol: ProtocolVersion,
    keypair: KeyPair<Ed25519>,
    profile: Profile,
    auth_nonce: [u8; 32],
    storage: BlockStore,
    invalid_block0_hash: bool,
}

Fields
genesis_hash: Hashprotocol: ProtocolVersionkeypair: KeyPair<Ed25519>profile: Profileauth_nonce: [u8; 32]storage: BlockStoreinvalid_block0_hash: bool
Implementations

impl MockServerData

const TIP_TAG: &'static str = "tip"

pub fn new( genesis_hash: Hash, protocol: ProtocolVersion, addr: SocketAddr, storage: BlockStore, invalid_get_blocks_hash: bool ) -> Self

pub fn genesis_hash(&self) -> &Hash

pub fn get_block(&self, header_id: Hash) -> Result<Block, Error>

pub fn genesis_block(&self) -> Block

pub fn tip(&self) -> Result<Header, Error>

pub fn set_tip(&self, block_id: &[u8]) -> Result<(), Error>

block_id must refer to a valid block already in the storage

pub fn put_block(&self, block: &Block) -> Result<(), Error>

pub fn profile(&self) -> &Profile

pub fn protocol(&self) -> &ProtocolVersion

pub fn node_id(&self) -> &[u8]

pub fn node_signature(&self, nonce: &[u8]) -> Vec<u8>

pub fn generate_auth_nonce(&mut self) -> &[u8]

pub fn storage(&self) -> &BlockStore

Raw access to the storage. Can be used to play with raw bytes sequences or get around validations of domain objects, for example to mimic a malicous adversary

pub fn validate_peer_node_id(&self, node_id: &[u8], signature: &[u8]) -> bool

pub fn genesis_hash_mut(&mut self) -> &mut Hash

pub fn protocol_mut(&mut self) -> &mut ProtocolVersion

pub fn invalid_block0_hash(&self) -> bool