Struct jormungandr_automation::jormungandr::RawRest

pub struct RawRest { /* private fields */ }

struct intends to return raw reqwest response can be used to verify requests error codes or to poll until data is available

Implementations

impl RawRest

pub fn new(uri: String, settings: RestSettings) -> Self

pub fn uri(&self) -> String

pub fn rest_settings(&self) -> &RestSettings

pub fn rest_settings_mut(&mut self) -> &mut RestSettings

pub fn enable_logger(&mut self)

pub fn disable_logger(&mut self)

pub fn epoch_reward_history(&self, epoch: u32) -> Result<Response, Error>

pub fn reward_history(&self, length: u32) -> Result<Response, Error>

pub fn remaining_rewards(&self) -> Result<Response, Error>

pub fn stake_distribution(&self) -> Result<Response, Error>

pub fn account_state(&self, id: &Identifier) -> Result<Response, Error>

pub fn account_state_by_pk(&self, bech32_str: &str) -> Result<Response, Error>

pub fn stake_pools(&self) -> Result<Response, Error>

pub fn account_votes(&self, address: Address) -> Result<Response, Error>

pub fn account_votes_with_plan_id(
&self,
vote_plan_id: VotePlanId,
address: Address
) -> Result<Response, Error>

pub fn stake_distribution_at(&self, epoch: u32) -> Result<Response, Error>

pub fn account_votes_all(&self) -> Result<Response, Error>

pub fn stats(&self) -> Result<Response, Error>

pub fn network_stats(&self) -> Result<Response, Error>

pub fn p2p_quarantined(&self) -> Result<Response, Error>

pub fn p2p_non_public(&self) -> Result<Response, Error>

pub fn p2p_available(&self) -> Result<Response, Error>

pub fn p2p_view(&self) -> Result<Response, Error>

pub fn leaders_log(&self) -> Result<Response, Error>

pub fn tip(&self) -> Result<Response, Error>

pub fn settings(&self) -> Result<Response, Error>

pub fn shutdown(&self) -> Result<Response, Error>

pub fn block(&self, header_hash: &HeaderId) -> Result<Response, Error>

pub fn fragment_logs(&self) -> Result<Response, Error>

pub fn send_fragment(&self, fragment: Fragment) -> Result<Response, Error>

pub fn send_raw_fragment(&self, body: Vec<u8>) -> Result<Response, Error>

pub fn send_raw_fragments(&self, bodies: Vec<Vec<u8>>) -> Result<(), Error>

pub fn fragments_logs(&self) -> Result<Response, Error>

pub fn fragments_statuses(&self, ids: Vec<String>) -> Result<Response, Error>

pub fn send_fragment_batch(
&self,
fragments: Vec<Fragment>,
fail_fast: bool
) -> Result<Response, Error>

pub fn vote_plan_statuses(&self) -> Result<Response, Error>

pub fn send_until_ok<F>(&self, action: F, wait: Wait) -> Result<(), RestError>where
F: Fn(&RawRest) -> Result<Response, Error>,